Added COMBAT SYSTEM!!! YAY! Added enemyID's in Locations, enemy parcing
Сделал комбат систему (РАБОЧУЮ!!!). Дополнил парсер локаций, добавив туда энеми айдишки, умом сделал удаление побежденных врагов с карты. Я все еще хз как это работает, но оно работает :D 😄

+++ Blade&Kami/Enemy.cpp
@@ -5,7 +5,7 @@ Enemy::Enemy(const EnemyData& baseData) : data(baseData) {
     data.currentHealth = data.maxHealth;
 }
 
-void Enemy::takeDamage(double damage) {
+void Enemy::takeDamage(int damage) {
     data.currentHealth -= damage;
     if (data.currentHealth < 0) data.currentHealth = 0;
 }
